backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic[email]==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.27.0
openai==1.50.0
supabase==2.7.0
bleach==6.1.0
